Added cast to fix warning

+++ include/Hdlcpp.hpp
@@ -228,7 +228,7 @@ private:
         if (!destination || !destinationLength)
             return -EINVAL;
 
-        for (i = 0; i < source.size(); i++) {
+        for (i = 0; i < static_cast<uint16_t>(source.size()); i++) {
             // First find the start flag sequence
             if (frameStartIndex < 0) {
                 if (source[i] == FlagSequence) {
